Changes to test-suite and non-working test for mappings
The test-suite now starts three cores instead of two. In the future, we
should make a programmatic way of starting arbitrary amount of cores.
Also the TCP ports for the cores have been changed to avoid clashes with
a "normal" core that might already be running on the host.

A test for mappings is added, but unfortunately it does not work with
the current setup - coersion works, but not the actual mapping operation